Add support for launcher boost for other launchers Supported Launchers: Pixel Launcher Nova Launcher OnePlus Launcher Lawnchair Launcher - Public CI/Alpha Builds
How does one determine the task names?
Use the last 15 characters of the package name, that's what the comm array stores. For an instance that the launcher package name contains less than 15 characters, use the full name.

However, this is useless for those launchers that do not use the new overview.
On OnePlus devices, I believe OnePlus uses their overview for ALL launchers, so all the above launchers will benefit from them. On other devices, only the stock launcher and above launchers will recieve this benefit, but only stock and Lawnchair(if Lawnstep is being used), the ones that currently support the new interface will recieve the benefit realistically.

Co-authored-by: Erik Müller <erik.lars.mueller@googlemail.com>
Signed-off-by: idkwhoiam322 <snkm1999@gmail.com>